General HTML, CSS, and JavaScript files
Added Icon image, description, and keyword meta tags to the head of the HTML file. Removed some inline styling for the bootstrap dropdown menu. Wrapped the h1 and table container in side of the main tag.  Formated the file to make it easier to read.

Changed the background color. Added animation for the dropdown menu and the list items inside. Formated the file to make it easier to read.

For the JavaScript file I formatted it to make it easier to read.